crt_sh/requirements.txt
psycopg2-binary
requests